+++ ecs/util.go
@@ -41,13 +41,6 @@ func copyPtr(src, dst unsafe.Pointer, itemSize uintptr) {
 	copy(dstSlice, srcSlice)
 }
 
-// copyValueReflect copies an item between two reflect arrays using reflection.
-// This is GC-safe. Use for non-trivial types.
-func copyValueReflect(src, dst reflect.Value, from, to int) {
-	// TODO: can potentially be optimized using typedmemmove
-	dst.Index(to).Set(src.Index(from))
-}
-
 // copyValue copies an item between two reflect arrays using typedmemmove.
 // This is GC-safe. Use for non-trivial types.
 func copyValue(src, dst *column, from, to uintptr) {